fix(build): Handle image bundling from css
`@rollup/plugin-url` only takes care of bundling images that are directly imported from ts/js
files and ignores images that are imported from css files. So adding this new postcss plugin
to handle bundling of images specified in css files.

Additionally, this ensures that we do not rely on postcss.config.js to be present in the directory.

NOTE: The `inline` option is chosen rather than `copy` since the latter expects the `assetPath`
to be provided and rewrites the `url(..)` patterns with that asset path. This is not ideal since
these assets will be moved over to a different location when bundled with a distribution like
`deck-nflx` and will result in 404s.