fix(build): Emit assets directly to dist folder
Having `[dirname]` specified in the output `filename` puts the asset in a new directory
named `[dirname]` under the dist folder. This gets copied over to the bundled output in
netflix's deck distribution, but doesn't get served correctly. In this commit, this is
changed so that the asset is emitted directly under dist folder.

The `limit` is also increased to allow inlining of images for plugins.