feat(core/pipeline): pipeline graph minimize based on number of stages
Pipeline graphs with many stages take a very long time to load. To
remediate this situation, pipeline graphs can now be minimized on
the initial mount of the pipeline graph based on a configurable
number of stages. When minimized, users can click 'Expand' to see
the full pipeline graph rendering.