fix(core/application): Fix for `requiredGroupMembership` breaking UI
Fixes a problem that can occur for some old apps that still have `requiredGroupMembership` set on the model in front50 (instead of or in addition to `permissions`). This breaks the rendering with a `$digest already in progress` error message. Removing the `$digest()` call and instead wrapping the code in an `$evalAsync()` callback solves the issue.